+++ DataCoreTests/Sql/TableScaleModels/PluV2ValidatorTests.cs
@@ -1,0 +1,90 @@
+﻿// This is an independent project of an individual developer. Dear PVS-Studio, please check it.
+// PVS-Studio Static Code Analyzer for C, C++, C#, and Java: http://www.viva64.com
+
+using DataCore.Sql.TableScaleModels;
+using FluentValidation.Results;
+using NSubstitute;
+using NUnit.Framework;
+using System.Linq;
+
+namespace DataCoreTests.Sql.TableScaleModels;
+
+[TestFixture]
+internal class PluV2ValidatorTests
+{
+	[Test]
+	public void Entity_Validate_IsFalse()
+	{
+		Assert.DoesNotThrow(() =>
+		{
+			// Arrange.
+			PluV2Entity item = Substitute.For<PluV2Entity>();
+			PluV2Validator validator = new();
+			// Act.
+			ValidationResult result = validator.Validate(item);
+			TestsUtils.FailureWriteLine(result);
+			// Assert.
+			Assert.IsFalse(result.IsValid);
+			// Act.
+			item.Name = "";
+			result = validator.Validate(item);
+			TestsUtils.FailureWriteLine(result);
+			// Assert.
+			Assert.IsFalse(result.IsValid);
+		});
+	}
+
+	[Test]
+	public void Entity_Validate_IsTrue()
+	{
+		Assert.DoesNotThrow(() =>
+		{
+			// Arrange.
+			PluV2Entity item = Substitute.For<PluV2Entity>();
+			PluV2Validator validator = new();
+			// Act.
+			item.Number = 100;
+			item.Name = "Test";
+			item.FullName = "Test";
+			item.Description = "Test";
+			item.Gtin = "Test";
+			item.Ean13 = "Test";
+			item.Itf14 = "Test";
+			ValidationResult result = validator.Validate(item);
+			TestsUtils.FailureWriteLine(result);
+			// Assert.
+			Assert.IsTrue(result.IsValid);
+		});
+	}
+
+	[Test]
+	public void DbTable_Validate_IsTrue()
+	{
+		TestsUtils.DbTableAction(() =>
+		{
+			// Arrange.
+			PluV2Validator validator = new();
+			PluV2Entity[]? items = TestsUtils.DataAccess.Crud.GetEntities<PluV2Entity>();
+			// Act.
+			if (items == null || !items.Any())
+			{
+				TestContext.WriteLine($"{nameof(items)} is null or empty!");
+			}
+			else
+			{
+				TestContext.WriteLine($"Found {nameof(items)}.Count: {items.Count()}");
+				int i = 0;
+				foreach (PluV2Entity item in items)
+				{
+					if (i < 10)
+						TestContext.WriteLine(item);
+					i++;
+					ValidationResult result = validator.Validate(item);
+					TestsUtils.FailureWriteLine(result);
+					// Assert.
+					Assert.IsTrue(result.IsValid);
+				}
+			}
+		});
+	}
+}